# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/channels/{id}": {
      "get": {
        "operationId": "ChannelsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ChannelResponseDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "ไม่พบห้อง หรือไม่ได้เป็นสมาชิก",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "not_found"
                        },
                        "message": {
                          "type": "string",
                          "example": "ไม่พบคลิปนี้ อาจถูกลบไปแล้ว"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "รายละเอียดห้อง (ต้องเป็นสมาชิก)",
        "tags": [
          "channels"
        ]
      }
    },
...
    "/api/v1/channels/{channelId}/messages": {
      "get": {
        "operationId": "MessagesController_list",
        "parameters": [
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "per_page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "after_seq",
            "required": false,
            "in": "query",
            "description": "ดึงเฉพาะข้อความที่ seq มากกว่าค่านี้ — ใช้ตอน socket หลุดแล้วต่อใหม่ เพื่อเติมช่วงที่ขาดโดยไม่ต้องโหลดทั้งห้อง",
            "schema": {
              "example": 1420,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "สำเร็จ พร้อมข้อมูลแบ่งหน้า",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MessageResponseDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "ไม่พบห้อง หรือไม่ได้เป็นสมาชิก",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "not_found"
                        },
                        "message": {
                          "type": "string",
                          "example": "ไม่พบคลิปนี้ อาจถูกลบไปแล้ว"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ประวัติข้อความในห้อง — ใส่ after_seq เพื่อเติมเฉพาะช่วงที่ขาดหลังต่อ socket ใหม่",
        "tags": [
          "messages"
        ]
      },
      "post": {
        "operationId": "MessagesController_send",
        "parameters": [
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/MessageResponseDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "ข้อความว่าง หรือไฟล์แนบยังไม่พร้อม",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "not_found"
                        },
                        "message": {
                          "type": "string",
                          "example": "ไม่พบคลิปนี้ อาจถูกลบไปแล้ว"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "แนบไฟล์ของคนอื่นไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "not_found"
                        },
                        "message": {
                          "type": "string",
                          "example": "ไม่พบคลิปนี้ อาจถูกลบไปแล้ว"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ส่งข้อความ (ทางเลือกสำรองของ socket)",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/v1/channels/{channelId}/messages/pinned": {
      "get": {
        "operationId": "MessagesController_listPinned",
        "parameters": [
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "per_page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "สำเร็จ พร้อมข้อมูลแบ่งหน้า",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MessageResponseDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ข้อความที่ปักหมุดในห้องนี้ — แผงประกาศของห้อง",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/v1/channels/{channelId}/messages/{messageId}/thread": {
      "get": {
        "operationId": "MessagesController_listThread",
        "parameters": [
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "per_page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 50,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "สำเร็จ พร้อมข้อมูลแบ่งหน้า",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/MessageResponseDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "ไม่พบข้อความต้นเธรด",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "not_found"
                        },
                        "message": {
                          "type": "string",
                          "example": "ไม่พบคลิปนี้ อาจถูกลบไปแล้ว"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "คำตอบทั้งหมดในเธรดของข้อความนี้ (เก่าไปใหม่)",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/v1/channels/{channelId}/messages/{messageId}": {
      "patch": {
        "description": "ผู้ดูแลห้องแก้ข้อความคนอื่นไม่ได้โดยตั้งใจ — ลบได้ แต่แก้ไม่ได้ เพราะการแก้คำพูดของคนอื่นแล้วยังแสดงชื่อเขาเป็นผู้เขียนคือการปลอมคำพูด",
        "operationId": "MessagesController_edit",
        "parameters": [
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditMessageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/MessageResponseDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "เกิน 15 นาทีแล้ว",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "not_found"
                        },
                        "message": {
                          "type": "string",
                          "example": "ไม่พบคลิปนี้ อาจถูกลบไปแล้ว"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "แก้ข้อความของคนอื่นไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "not_found"
                        },
                        "message": {
                          "type": "string",
                          "example": "ไม่พบคลิปนี้ อาจถูกลบไปแล้ว"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "แก้ข้อความของตัวเอง ภายใน 15 นาทีหลังส่ง",
        "tags": [
          "messages"
        ]
      },
      "delete": {
        "operationId": "MessagesController_remove",
        "parameters": [
          {
            "name": "channelId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "messageId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "ลบข้อความของคนอื่นไม่ได้",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "not_found"
                        },
                        "message": {
                          "type": "string",
                          "example": "ไม่พบคลิปนี้ อาจถูกลบไปแล้ว"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ลบข้อความ (เจ้าของ ผู้ดูแลห้อง หรือ admin องค์กร)",
        "tags": [
          "messages"
        ]
      }
    },
...
  },
  "info": {
    "title": "CS Nexus — aie4-social-reels",
    "description": "ระบบย่อยภายใต้ CSMJU2030 · ตัวตนมาจาก API Gateway ผ่าน header X-User-Id / X-Layer1-Role / X-Faculty",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "current_page": {
            "type": "number",
            "example": 2
          },
          "per_page": {
            "type": "number",
            "example": 50
          },
          "total_pages": {
            "type": "number",
            "example": 50
          },
          "total_items": {
            "type": "number",
            "example": 2480
          }
        },
        "required": [
          "current_page",
          "per_page",
          "total_pages",
          "total_items"
        ]
      },
...
      "ChannelResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "DM",
              "GROUP",
              "COURSE",
              "VOICE"
            ]
          },
          "name": {
            "type": "object",
            "nullable": true
          },
          "course_tag": {
            "type": "object",
            "nullable": true
          },
          "max_seats": {
            "type": "number"
          },
          "member_count": {
            "type": "number"
          },
          "my_role": {
            "type": "string",
            "enum": [
              "MEMBER",
              "MODERATOR"
            ]
          },
          "unread_count": {
            "type": "number",
            "description": "จำนวนข้อความที่ยังไม่ได้อ่าน"
          },
          "created_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "kind",
          "name",
          "course_tag",
          "max_seats",
          "member_count",
          "my_role",
          "unread_count",
          "created_at"
        ]
      },
...
      "MessageAttachmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "file_name": {
            "type": "string"
          },
          "kind": {
            "type": "string"
          },
          "mime_type": {
            "type": "string"
          },
          "size_bytes": {
            "type": "string",
            "example": "48120"
          }
        },
        "required": [
          "id",
          "file_name",
          "kind",
          "mime_type",
          "size_bytes"
        ]
      },
      "MessageEmbedDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "REEL",
              "POST"
            ]
          },
          "ref_id": {
            "type": "string"
          }
        },
        "required": [
          "kind",
          "ref_id"
        ]
      },
      "MessageResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "seq": {
            "type": "number",
            "description": "ลำดับจริงของข้อความในห้อง — ใช้ตัวนี้เรียง ไม่ใช่ created_at",
            "example": 1421
          },
          "channel_id": {
            "type": "string"
          },
          "author_username": {
            "type": "string"
          },
          "content": {
            "type": "object",
            "nullable": true
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MessageAttachmentDto"
            }
          },
          "embed": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/MessageEmbedDto"
              }
            ]
          },
          "parent_id": {
            "type": "object",
            "nullable": true,
            "description": "ถ้าไม่ null ข้อความนี้เป็นคำตอบในเธรดของ id นั้น"
          },
          "reply_count": {
            "type": "number",
            "example": 0,
            "description": "จำนวนคำตอบในเธรดของข้อความนี้"
          },
          "pinned_at": {
            "type": "object",
            "nullable": true
          },
          "pinned_by_username": {
            "type": "object",
            "nullable": true
          },
          "client_nonce": {
            "type": "string"
          },
          "edited_at": {
            "type": "object",
            "nullable": true
          },
          "created_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "seq",
          "channel_id",
          "author_username",
          "content",
          "attachments",
          "embed",
          "parent_id",
          "reply_count",
          "pinned_at",
          "pinned_by_username",
          "client_nonce",
          "edited_at",
          "created_at"
        ]
      },
      "SendMessageDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "example": "ส่งไฟล์โปรเจกต์ให้แล้วนะ"
          },
          "asset_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "embed": {
            "description": "แชร์คลิป Reels หรือโพสต์เข้าแชท — แสดงเป็น mini player",
            "allOf": [
              {
                "$ref": "#/components/schemas/MessageEmbedDto"
              }
            ]
          },
          "parent_id": {
            "type": "string",
            "description": "ตอบกลับในเธรดของข้อความนี้ (แบบ Discord/Teams) — ข้อความที่มี parent_id จะไม่โผล่ในไทม์ไลน์หลัก และเธรดซ้อนเธรดไม่ได้"
          },
          "client_nonce": {
            "type": "string",
            "description": "client สร้างเอง กันส่งซ้ำตอนเน็ตกระตุก และใช้เป็นคีย์ชั่วคราวใน UI",
            "example": "01J9F2K7Q8"
          }
        },
        "required": [
          "client_nonce"
        ]
      },
      "EditMessageDto": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string",
            "example": "ขอแก้เป็น struct ไม่ใช่ class ครับ"
          }
        },
        "required": [
          "content"
        ]
      },
...
    }
  }
}